IN-183-my-payments: add useMyPayments
